+++ src/Router/RouteInterface.php
@@ -6,5 +6,5 @@ use Psr\Http\Server\MiddlewareInterface;
 
 interface RouteInterface
 {
-    public function middleware(MiddlewareInterface $middleware): self;
+
 }